requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
requests>=2.31.0
